+++ GBA_emu/Utils.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 
-uint32_t rotateRight(uint32_t word, uint8_t n);
+uint32_t rotateRight(uint32_t word, uint32_t n);
 
-uint32_t arithmeticShift(uint32_t word, uint8_t n);
+uint32_t arithmeticShift(uint32_t word, uint32_t n);
